feat: Starting the tests

+++ backend/src/main/java/com/example/bluebird/api/repository/SignUpRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.bluebird.api.models.SignUpModel;
 
-public interface SignUpRepository extends JpaRepository<SignUpModel, Long> {
+public interface SignUpRepository extends JpaRepository<SignUpModel, Integer> {
     
 }
